+++ kabac/src/parser/function.rs
@@ -2,8 +2,7 @@ use super::{
     block,
     error::{ParsingError, ParsingErrorVariant},
     state::ParserState,
-    tn::TypeNotationParser,
-    Result,
+    tn, Result,
 };
 use crate::{
     ast::{AstNode, FunctionParam},
@@ -14,20 +13,23 @@ pub fn parse(state: &ParserState) -> Result<AstNode> {
     let start = state.tokens.current().span.start;
 
     // Expecting "fn" keyword
-    state.tokens.skip(&TokenKind::Fn)?;
+    state.tokens.skip(&TokenKind::Def)?;
 
     // Expecting symbol
     let sym_id = state.next_symbol_id();
     let sym = parse_sym(state)?;
 
-    // Expecting "("
-    state.tokens.skip(&TokenKind::LParen)?;
+    // Expecting parameters (optional)
+    let params = if state.tokens.current_is(&TokenKind::LParen) {
+        // Expecting >= 0 function parameter declaration(s)
+        state.tokens.skip(&TokenKind::LParen)?;
+        let params = parse_params(state)?;
+        state.tokens.skip(&TokenKind::RParen)?;
 
-    // Expecting >= 0 function parameter declaration(s)
-    let params = parse_params(state)?;
-
-    // Expecting ")"
-    state.tokens.skip(&TokenKind::RParen)?;
+        params
+    } else {
+        vec![]
+    };
 
     // Expecting return type notation (optional)
     let return_tn = parse_return_tn(state)?;
@@ -79,7 +81,7 @@ fn parse_params(state: &ParserState) -> Result<Vec<FunctionParam>> {
         state.tokens.skip(&TokenKind::Colon)?;
 
         // Expecting type notation
-        let tn = TypeNotationParser::new(state).parse()?;
+        let tn = tn::parse(state)?;
 
         params.push(FunctionParam { sym, sym_id, tn });
 
@@ -115,7 +117,7 @@ fn parse_return_tn(state: &ParserState) -> Result<Option<AstNode>> {
 
     state.tokens.skip(&TokenKind::Colon)?;
 
-    Ok(Some(TypeNotationParser::new(state).parse()?))
+    Ok(Some(tn::parse(state)?))
 }
 
 #[cfg(test)]
@@ -128,18 +130,37 @@ mod tests {
     #[test]
     fn empty_function_definition() {
         assert_ast(
-            "fn foo() {}",
+            "def foo {}",
             AstNode::FunctionDefinition {
                 sym: Box::new(AstNode::Symbol {
                     name: String::from("foo"),
-                    span: 3..6,
+                    span: 4..7,
                 }),
                 sym_id: 1,
                 params: vec![],
                 return_tn: None,
                 body: vec![],
                 scope_id: 2,
-                span: 0..11,
+                span: 0..10,
+            },
+        );
+    }
+
+    #[test]
+    fn function_definition_with_zero_parameters_but_with_parentheses() {
+        assert_ast(
+            "def foo() {}",
+            AstNode::FunctionDefinition {
+                sym: Box::new(AstNode::Symbol {
+                    name: String::from("foo"),
+                    span: 4..7,
+                }),
+                sym_id: 1,
+                params: vec![],
+                return_tn: None,
+                body: vec![],
+                scope_id: 2,
+                span: 0..12,
             },
         );
     }
@@ -147,41 +168,41 @@ mod tests {
     #[test]
     fn function_definition_with_parameters_and_trailing_comma() {
         assert_ast(
-            "fn foo(x: int, y: bool,) {}",
+            "def foo(x: int, y: bool,) {}",
             AstNode::FunctionDefinition {
                 sym: Box::new(AstNode::Symbol {
                     name: String::from("foo"),
-                    span: 3..6,
+                    span: 4..7,
                 }),
                 sym_id: 1,
                 params: vec![
                     FunctionParam {
                         sym: AstNode::Symbol {
                             name: String::from("x"),
-                            span: 7..8,
+                            span: 8..9,
                         },
                         sym_id: 2,
                         tn: AstNode::TypeNotation {
                             tn: TypeNotation::Symbol(String::from("int")),
-                            span: 10..13,
+                            span: 11..14,
                         },
                     },
                     FunctionParam {
                         sym: AstNode::Symbol {
                             name: String::from("y"),
-                            span: 15..16,
+                            span: 16..17,
                         },
                         sym_id: 3,
                         tn: AstNode::TypeNotation {
                             tn: TypeNotation::Symbol(String::from("bool")),
-                            span: 18..22,
+                            span: 19..23,
                         },
                     },
                 ],
                 return_tn: None,
                 body: vec![],
                 scope_id: 2,
-                span: 0..27,
+                span: 0..28,
             },
         );
     }
@@ -189,38 +210,38 @@ mod tests {
     #[test]
     fn function_definition_with_parameter_and_body() {
         assert_ast(
-            "fn write(x: int) { print(x); }",
+            "def write(x: int) { print(x); }",
             AstNode::FunctionDefinition {
                 sym: Box::new(AstNode::Symbol {
                     name: String::from("write"),
-                    span: 3..8,
+                    span: 4..9,
                 }),
                 sym_id: 1,
                 params: vec![FunctionParam {
                     sym: AstNode::Symbol {
                         name: String::from("x"),
-                        span: 9..10,
+                        span: 10..11,
                     },
                     sym_id: 2,
                     tn: AstNode::TypeNotation {
                         tn: TypeNotation::Symbol(String::from("int")),
-                        span: 12..15,
+                        span: 13..16,
                     },
                 }],
                 return_tn: None,
                 body: vec![AstNode::FunctionCall {
                     callee: Box::new(AstNode::Symbol {
                         name: String::from("print"),
-                        span: 19..24,
+                        span: 20..25,
                     }),
                     args: vec![AstNode::Symbol {
                         name: String::from("x"),
-                        span: 25..26,
+                        span: 26..27,
                     }],
-                    span: 19..27,
+                    span: 20..28,
                 }],
                 scope_id: 2,
-                span: 0..30,
+                span: 0..31,
             },
         );
     }
@@ -228,27 +249,27 @@ mod tests {
     #[test]
     fn function_definition_with_return_statement() {
         assert_ast(
-            "fn foo(): int { return 5; }",
+            "def foo: int { return 5; }",
             AstNode::FunctionDefinition {
                 sym: Box::new(AstNode::Symbol {
                     name: String::from("foo"),
-                    span: 3..6,
+                    span: 4..7,
                 }),
                 sym_id: 1,
                 params: vec![],
                 return_tn: Some(Box::new(AstNode::TypeNotation {
                     tn: TypeNotation::Symbol(String::from("int")),
-                    span: 10..13,
+                    span: 9..12,
                 })),
                 body: vec![AstNode::Return {
                     expr: Some(Box::new(AstNode::Literal {
                         lit: Literal::Int(5),
-                        span: 23..24,
+                        span: 22..23,
                     })),
-                    span: 16..24,
+                    span: 15..23,
                 }],
                 scope_id: 2,
-                span: 0..27,
+                span: 0..26,
             },
         );
     }