+++ kabac/src/semantic/each_loop.rs
@@ -118,7 +118,7 @@ mod tests {
     #[test]
     fn each_loop_statements() {
         assert_is_ok(indoc! {"
-                fn main() {
+                def main {
                     each n in [int 1, 2, 3] {
                         debug n;
                     }
@@ -134,7 +134,7 @@ mod tests {
     #[test]
     fn each_loop_statement_with_non_array_expression() {
         assert_is_err(indoc! {"
-                fn main() {
+                def main {
                     each n in true {
                         debug n;
                     }
@@ -145,7 +145,7 @@ mod tests {
     #[test]
     fn accessing_elem_id_outside_scope() {
         assert_is_err(indoc! {"
-                fn main() {
+                def main {
                     each n in [int 1, 2] {}
 
                     debug n;