+++ kabac/src/semantic/while_loop.rs
@@ -92,7 +92,7 @@ mod tests {
     #[test]
     fn while_loop_statements() {
         assert_is_ok(indoc! {"
-                fn main() {
+                def main {
                     while 2 > 5 {
                         debug 1;
                     }
@@ -111,7 +111,7 @@ mod tests {
     #[test]
     fn using_math_expression_as_condition_in_while_statement() {
         assert_is_err(indoc! {"
-                fn main() {
+                def main {
                     while 5 + 5 {}
                 }
             "})
@@ -120,7 +120,7 @@ mod tests {
     #[test]
     fn using_break_statement_not_in_loop_scope() {
         assert_is_err(indoc! {"
-                fn main() {
+                def main {
                     if true {
                         break;
                     }
@@ -131,7 +131,7 @@ mod tests {
     #[test]
     fn using_invalid_statement_after_loop_control() {
         assert_is_err(indoc! {"
-                fn main() {
+                def main {
                     while true {
                         break;
                         1 + true; // this should be error