+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- `fn` keyword is replaced with `def` keyword.
+- Make parentheses optional in function definition when it does not receive any parameters.
+
 ## [0.5.3] - 2025-02-03
 
 ### Fixed